feat: cleanup and documentation for aggregate functions
