--- src/main/java/de/featjar/formula/structure/term/aggregate/IAggregate.java
@@ -1,6 +1,0 @@
-package de.featjar.formula.structure.term.aggregate;
-
-import de.featjar.formula.structure.term.ITerm;
-
-public interface IAggregate extends ITerm {
-}
