+++ src/main/java/de/featjar/formula/structure/term/aggregate/IAttributeAggregate.java
@@ -1,0 +1,14 @@
+package de.featjar.formula.structure.term.aggregate;
+
+import de.featjar.formula.structure.term.ITerm;
+
+/**
+ * Interface for modelling attribute aggregate functionality. Currently, attributes aren't modeled in first order
+ * predicate logic. Therefore, this is more like a placeholder then a mathematically correct implementation.
+ *
+ * @author Lara Merza
+ * @author Felix
+ * @author Jonas Hanke
+ */
+public interface IAttributeAggregate extends ITerm {
+}
